chore(restructure):directory restructure - restructure directories to remove folder clut - resolves #35
