+++ src/controllers/authController.ts
@@ -1,14 +1,14 @@
 import bcrypt from "bcrypt";
 import { Request, Response, NextFunction, Router } from "express";
-import Controller from "../interfaces/controller.interface";
-import validationMiddleware from "../middleware/validation.middleware";
+import Controller from "../interfaces/controllerInterface";
+import validationMiddleware from "../middleware/valdationMiddleware";
 import UserWithThatEmailAlreadyExistsException from "../exceptions/UserWithThatEmailAlreadyExistsException";
-import LogInDto from "./login.dto";
+import LogInDto from "../dtos/loginDto";
 import WrongCredentialsException from "../exceptions/WrongCredentialsException";
-import CreateUserDto from "../users/user.dto";
+import CreateUserDto from "../dtos/userDto";
 import jwt from "jsonwebtoken";
-import TokenData from "../interfaces/tokenData.interface";
-import DataStoredInToken from "../interfaces/dataStoredInToken";
+import TokenData from "../interfaces/tokenDataInterface";
+import DataStoredInToken from "../interfaces/dataStoredInTokenInterface";
 import { config } from "../Utils/config";
 import { V2_BASE_URL } from "../Utils/constants";
 import { getRepository } from "typeorm";
