+++ src/interfaces/dataStoredInTokenInterface.ts
