+++ src/interfaces/postInterface.ts
