+++ src/interfaces/requestWithUserInterface.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import User from "./user.interface";
+import User from "./userInterface";
 
 interface RequestWithUser extends Request {
   user: User;
