+++ src/interfaces/tokenDataInterface.ts
