+++ src/middleware/authMiddleware.ts
@@ -2,8 +2,8 @@ import { Response, NextFunction } from "express";
 import jwt from "jsonwebtoken";
 import AuthenticationTokenMissingException from "../exceptions/AuthenticationTokenMissingException";
 import WrongAuthenticationTokenException from "../exceptions/WrongAuhenticationTokenException";
-import DataStoredInToken from "../interfaces/dataStoredInToken";
-import RequestWithUser from "../interfaces/requestWithUser.interface";
+import DataStoredInToken from "../interfaces/dataStoredInTokenInterface";
+import RequestWithUser from "../interfaces/requestWithUserInterface";
 import { config } from "../Utils/config";
 import ExpiredAuthenticationTokenException from "../exceptions/ExpiredAuthenticationTokenException";
 import User from "../entity/user";
