--- src/users/user.models.ts
@@ -1,19 +1,0 @@
-// import { Schema, model, Document } from "mongoose";
-// import User from "./user.interface";
-
-// const addressSchema = new Schema({
-//   city: String,
-//   street: String,
-//   country: String
-// });
-
-// const userSchema = new Schema({
-//   name: String,
-//   address: addressSchema,
-//   email: String,
-//   password: String,
-//   posts: [{ ref: "Post", type: Schema.Types.ObjectId }]
-// });
-
-// const userModel = model<User & Document>("User", userSchema);
-// export default userModel;
